Changes: - Revert to Scala LTS version `3.3.5` - Simplify `scalafixAll` and `scalafmt` - Merge modules `basic`, `standard`, `advanced` and `magic` into one module `language-feature` - Merge `types`, `domain`, `errors` and `Main` under one object - Upgrade `neotype` to version `0.3.11` - Update README.md hyperlinks to source code.
