+++ .scalafix.conf
@@ -3,11 +3,10 @@ rules = [
 ]
 
 OrganizeImports {
-  groupedImports = Explode
+  expandRelative = true
   groups = [
     "re:(java?|scala)\\.",
     "dagmendez",
     "*"
   ]
-  removeUnused = false
 }
