+++ .scalafmt.conf
@@ -2,9 +2,10 @@ version = "3.8.6"
 
 maxColumn = 150
 
-runner {
-  dialect = scala3
-}
+align.preset = more
+docstrings.style = Asterisk
+rewrite.rules = [RedundantBraces]
+runner.dialect = scala3
 
 fileOverride {
   "glob:**.sbt" {
@@ -13,23 +14,4 @@ fileOverride {
   "glob:**/project/**.scala" {
     runner.dialect = scala213
   }
-}
-
-align {
-  arrowEnumeratorGenerator = true
-  ifWhileOpenParen = false
-  openParenCallSite = true
-  openParenDefnSite = true
-  preset = more
-}
-
-docstrings.style = Asterisk
-
-rewrite {
-  rules = [RedundantBraces]
-  redundantBraces.maxLines = 1
-}
-
-optIn {
-  breaksInsideChains = true
 }
