--- 01-basic/src/main/scala/dagmendez/basic/domain.scala
@@ -1,10 +1,0 @@
-package dagmendez.basic
-
-import dagmendez.basic.typeAliases._
-
-object domain:
-
-  // The account holder is the person who signs the contract for said account with the bank
-  final case class AccountHolder(firstName: Name, middleName: Option[Name], lastName: Name, secondLastName: Option[Name])
-
-  final case class Account(accountHolder: AccountHolder, iban: IBAN, balance: Balance)
