--- 01-basic/src/main/scala/dagmendez/basic/typeAliases.scala
@@ -1,7 +1,0 @@
-package dagmendez.basic
-
-object typeAliases:
-
-  type Name    = String
-  type IBAN    = String // International Bank Account Number
-  type Balance = Int
