--- 03-advanced/src/main/scala/dagmendez/advanced/Main.scala
@@ -1,99 +1,0 @@
-package dagmendez.advanced
-
-import scala.util.control.NoStackTrace
-
-import dagmendez.advanced.domain._
-import dagmendez.advanced.error._
-import dagmendez.advanced.opaqueTypes._
-
-object Main:
-
-  private object HappyApply:
-    private val firstName: Name  = Name("John")
-    private val middleName: Name = Name("Stuart")
-    private val lastName: Name   = Name("Mill")
-    private val iban: IBAN       = IBAN("GB33BUKB20201555555555")
-    private val balance: Balance = Balance(-300)
-
-    private val account: Account =
-      Account(
-        AccountHolder(
-          firstName,
-          Some(middleName),
-          lastName,
-          secondLastName = None
-        ),
-        iban,
-        balance
-      )
-
-    def print(): Unit = println(account)
-
-  private object UnhappyApply:
-    private val firstName: Name = Name("John") // Comment this one an uncomment next line
-    // private val firstName: Name = Name("") // Uncomment and won't compile
-    private val middleName: Name = Name("Stuart")
-    private val lastName: Name   = Name("Mill")
-    private val iban: IBAN       = IBAN("GB33BUKB20201555555555")
-    private val balance: Balance = Balance(-1000)
-
-    private val account: Account =
-      Account(
-        AccountHolder(
-          firstName,
-          Some(middleName),
-          lastName,
-          secondLastName = None
-        ),
-        iban,
-        balance
-      )
-
-    def print(): Unit = println(account)
-
-  private object HappyFrom:
-    private val firstName: Either[InvalidName, Name]     = Name.either("John")
-    private val middleName: Either[InvalidName, Name]    = Name.either("Stuart")
-    private val lastName: Either[InvalidName, Name]      = Name.either("Mill")
-    private val iban: Either[InvalidIBAN, IBAN]          = IBAN.either("GB33BUKB20201555555555")
-    private val balance: Either[InvalidBalance, Balance] = Balance.either(0)
-
-    private val account: Either[RuntimeException with NoStackTrace, Account] =
-      for
-        fn <- firstName
-        mn <- middleName
-        ln <- lastName
-        ib <- iban
-        bl <- balance
-      yield Account(AccountHolder(fn, Some(mn), ln, secondLastName = None), ib, bl)
-
-    assert(account.isRight)
-    def print(): Unit = println(account)
-
-  private object UnhappyFrom:
-
-    // Play with any field that would crash the validation and return Left
-    private val firstName: Either[InvalidName, Name]     = Name.either("John")
-    private val middleName: Either[InvalidName, Name]    = Name.either("Stuart ") // This returns Left.
-    private val lastName: Either[InvalidName, Name]      = Name.either("Mill")
-    private val iban: Either[InvalidIBAN, IBAN]          = IBAN.either("GB33BUKB20201555555555")
-    private val balance: Either[InvalidBalance, Balance] = Balance.either(-5000)  // This returns Left.
-
-    private val account: Either[RuntimeException with NoStackTrace, Account] =
-      for
-        fn <- firstName
-        mn <- middleName
-        ln <- lastName
-        ib <- iban
-        bl <- balance
-      yield Account(AccountHolder(fn, Some(mn), ln, secondLastName = None), ib, bl)
-
-    assert(account.isLeft)
-    def print(): Unit = println(account)
-
-  @main def run(): Unit = {
-    HappyApply.print()   // Compiles
-    UnhappyApply.print() // Won't compile
-    HappyFrom.print()    // Right
-    UnhappyFrom.print()  // Left
-  }
