--- 03-advanced/src/main/scala/dagmendez/advanced/domain.scala
@@ -1,9 +1,0 @@
-package dagmendez.advanced
-
-import dagmendez.advanced.opaqueTypes._
-object domain:
-
-  // The account holder is the person who signs the contract for said account with the bank
-  final case class AccountHolder(firstName: Name, middleName: Option[Name], lastName: Name, secondLastName: Option[Name])
-
-  final case class Account(accountHolder: AccountHolder, iban: IBAN, balance: Balance)
