--- 04-scala-magic/src/main/scala/dagmendez/magic/error.scala
@@ -1,11 +1,0 @@
-package dagmendez.magic
-
-import scala.util.control.NoStackTrace
-
-object error:
-
-  final case class InvalidName(message: String) extends RuntimeException(message) with NoStackTrace
-
-  final case class InvalidIBAN(message: String) extends RuntimeException(message) with NoStackTrace
-
-  final case class InvalidBalance(message: String) extends RuntimeException(message) with NoStackTrace
