+++ build.sbt
@@ -1,55 +1,34 @@
+
 ThisBuild / organization := "dagmendez"
 
-ThisBuild / scalaVersion := "3.6.3"
+ThisBuild / scalaVersion := "3.3.5"
 
 ThisBuild / semanticdbEnabled := true
 ThisBuild / semanticdbVersion := scalafixSemanticdb.revision
+
+ThisBuild / scalacOptions ++= Seq("-Wunused:all")
+
 
 lazy val `opaque-types-and-inline`: Project =
   project
     .in(file("."))
     .aggregate(
-      basic,
-      standard,
-      advanced,
-      `scala-magic`,
+      `language-feature`,
       workshop,
       iron,
       neotype
     )
 
-lazy val basic: Project =
+lazy val `language-feature`: Project =
   project
-    .in(file("01-basic"))
-    .settings(commonScalacOptions)
-    .settings(name := "basic")
-
-lazy val standard: Project =
-  project
-    .in(file("02-standard"))
-    .settings(commonScalacOptions)
-    .settings(name := "standard")
-
-lazy val advanced: Project =
-  project
-    .in(file("03-advanced"))
-    .settings(commonScalacOptions)
-    .settings(name := "advanced")
-
-lazy val `scala-magic`: Project =
-  project
-    .in(file("04-scala-magic"))
-    .settings(commonScalacOptions)
-    .settings(
-      name := "scala-magic",
-      libraryDependencies ++= Seq(
-      )
-    )
+    .in(file("01-language-feature"))
+    .settings(commonSettings)
+    .settings(name := "language-feature")
 
 lazy val `workshop`: Project = {
   project
-    .in(file("06-workshop"))
-    .settings(commonScalacOptions)
+    .in(file("02-workshop"))
+    .settings(commonSettings)
     .settings(
       name := "workshop",
       libraryDependencies ++= Seq(
@@ -59,8 +38,8 @@ lazy val `workshop`: Project = {
 
 lazy val iron: Project = {
   project
-    .in(file("07-iron"))
-    .settings(commonScalacOptions)
+    .in(file("04-iron"))
+    .settings(commonSettings)
     .settings(
       name := "iron",
       libraryDependencies ++= Seq(
@@ -71,20 +50,22 @@ lazy val iron: Project = {
 
 lazy val neotype: Project = {
   project
-    .in(file("08-neotype"))
-    .settings(commonScalacOptions)
+    .in(file("03-neotype"))
+    .settings(commonSettings)
     .settings(
       name := "neotype",
       libraryDependencies ++= Seq(
-        "io.github.kitlangton" %% "neotype" % "0.3.10"
+        "io.github.kitlangton" %% "neotype" % "0.3.11"
       )
     )
 }
 
+lazy val commonSettings = commonScalacOptions ++ Seq(resolvers += "confluent" at "https://packages.confluent.io/maven/")
+
 lazy val commonScalacOptions = Seq(
   Compile / console / scalacOptions --= Seq(
     "-Xfatal-warnings",
-    "-Wunused:imports"
+    "-Wunused:_"
   ),
   Test / console / scalacOptions := (Compile / console / scalacOptions).value
 )
